Error Handling

+++ Application/Activities/Delete.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using Application.Core;
+using MediatR;
 using Persistence;
 using System;
 using System.Collections.Generic;
@@ -11,12 +12,12 @@ namespace Application.Activities
 {
     public  class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private DataContext _context;
 
@@ -25,13 +26,16 @@ namespace Application.Activities
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var deletedActivity = await _context.Activities.FindAsync(request.Id);
+                if (deletedActivity == null) 
+                    return null;
                 _context.Activities.Remove(deletedActivity);
-                await _context.SaveChangesAsync();
-
-                return Unit.Value;
+                var result= await _context.SaveChangesAsync() > 0;
+                if (!result)
+                    return Result<Unit>.Failure("Failed to delete the activity");
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
